requirements.txt
aiogram==3.0.0b7
openai==0.26.5
python-dotenv==1.0.0